+++ extensions/google-workspace/src/helpers/docs.ts
@@ -1,9 +1,14 @@
-import { open } from "@raycast/api";
+import { open, getPreferenceValues } from "@raycast/api";
 import { getUserEmail } from "../api/googleAuth";
 
 export type Prefix = "document" | "forms" | "spreadsheets" | "presentation";
 
+interface Preferences {
+  preferredBrowser: string;
+}
+
 export async function createDocFromUrl(prefix: Prefix, title?: string) {
+  const { preferredBrowser } = getPreferenceValues<Preferences>();
   const email = await getUserEmail();
 
   const baseUrl = `https://docs.google.com/${prefix}/create`;
@@ -18,5 +23,5 @@ export async function createDocFromUrl(prefix: Prefix, title?: string) {
 
   const url = baseUrl + "?" + searchParams.toString();
 
-  await open(url);
+  await open(url, preferredBrowser || undefined);
 }